requirements.txt
requests~=2.23.0
bs4~=0.0.1
beautifulsoup4~=4.9.1
python-wordpress-xmlrpc~=2.3
